EventBuilder - Add bookmark for struct subfield count
Allow updating a struct's sub-field count after the struct has been
started. This enables support for cases where we don't know the number
of sub-fields that a struct will have at the time the struct is started,
i.e. if we're adding a sub-field for each item in a collection, but the
collection doesn't have a `len()` method.

Note that this does NOT enable structs with 0 sub-fields (a struct MUST
have at least one sub-field and cannot have more than 127 sub-fields)
and it does NOT enable adding fields out-of-order (you still need to add
all the fields to the current struct before you can start a new struct).

Fixes #5