eventheader - check absolute paths for user_events_data
Previously, the trace provider initialization code located the user_events_data file by parsing `/proc/mounts` to find the tracefs or debugfs mount point, then used the mount point to form the full path.

This works fairly well, but it is an obstacle for container scenarios - the container creator probably doesn't want to project the entire `tracefs` or `debugfs` filesystem into the container, but they have to fake it by having a `/proc/mounts` file that points at a fake tracing directory.

Fix by looking for `user_events_data` in two fixed locations before trying to parse `/proc/mounts`. This means a container creator only has to project a `user_events_data` file and doesn't have to create a dummy `/proc/mounts` file. It will also tend to be more efficient for non-container scenarios where the `user_events_data` file is present (if present, it will almost always be in one of the probed paths). The downside is that in cases where it isn't present, we now probe two additional locations before giving up.

(Alternative: only look in `/sys/kernel/tracing`, don't look in `/sys/kernel/debug/tracing`. That would be more efficient in the case where the file does not exist, and would still achieve the goal of enabling containers, but it would be less efficient in a somewhat-common case where only debugfs is mounted.)