eventheader - check absolute paths for user_events_data (#8)
Previously, the trace provider initialization code located the user_events_data file by parsing `/proc/mounts` to find the tracefs or debugfs mount point, then used the mount point to form the full path.

This works fairly well, but it is an obstacle for container scenarios - the container creator probably doesn't want to mount the entire `tracefs` or `debugfs` filesystem into the container, so they have to fake it by having a `/proc/mounts` file that points at a fake tracing directory.

Fix by looking for `user_events_data` in fixed path `/sys/kernel/tracing` before trying to parse `/proc/mounts`. This means a container creator only has to project a `user_events_data` file and doesn't have to create a dummy `/proc/mounts` file. It will also tend to be more efficient for non-container scenarios where the `user_events_data` file is present (if present, it will almost always be in the probed path). The downside is that in cases where it isn't present, we now probe an additional location before giving up.